Revery moving of functions to reduce diff size
