+++ assets/js/base/components/cart-checkout/address-form/validate-shipping-country.ts
@@ -2,48 +2,13 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
-import type { ShippingAddress } from '@woocommerce/settings';
+import { type ShippingAddress } from '@woocommerce/settings';
 import { select, dispatch } from '@wordpress/data';
 import { VALIDATION_STORE_KEY } from '@woocommerce/block-data';
-import { isPostcode } from '@woocommerce/blocks-checkout';
-
-/**
- * Custom validation handler for fields with field specific handling.
- */
-export const customValidationHandler = (
-	inputObject: HTMLInputElement,
-	field: string,
-	customValues: {
-		country: string;
-	}
-): boolean => {
-	// Pass validation if the field is not required and is empty.
-	if ( ! inputObject.required && ! inputObject.value ) {
-		return true;
-	}
-
-	if (
-		field === 'postcode' &&
-		customValues.country &&
-		! isPostcode( {
-			postcode: inputObject.value,
-			country: customValues.country,
-		} )
-	) {
-		inputObject.setCustomValidity(
-			__(
-				'Please enter a valid postcode',
-				'woo-gutenberg-products-block'
-			)
-		);
-		return false;
-	}
-	return true;
-};
 
 // If it's the shipping address form and the user starts entering address
 // values without having set the country first, show an error.
-export const validateShippingCountry = ( values: ShippingAddress ): void => {
+const validateShippingCountry = ( values: ShippingAddress ): void => {
 	const validationErrorId = 'shipping_country';
 	const hasValidationError =
 		select( VALIDATION_STORE_KEY ).getValidationError( validationErrorId );
@@ -74,3 +39,5 @@ export const validateShippingCountry = ( values: ShippingAddress ): void => {
 		);
 	}
 };
+
+export default validateShippingCountry;
